refactor(playground): extract compare-pane feature
Move pane lifecycle and rendering out of main.ts:

features/compare-pane/
  - pane.ts: Pane interface, makePane, disposePane, forEachPane
  - handles.ts: PaneHandles interface
  - render.ts: renderPane (metrics push + bubble evict + draw)
  - bubbles.ts: updateBubbles, resolveStopName
  - decision.ts: pushDecision + decay logic
  - mode-badge.ts: updateModeBadge

main.ts drops from 1038 to 776 lines.
