+++ playground/src/features/compare-pane/index.ts
@@ -1,0 +1,14 @@
+export { type PaneHandles } from "./handles";
+export {
+  type Pane,
+  type PaneState,
+  COLOR_A,
+  COLOR_B,
+  makePane,
+  disposePane,
+  forEachPane,
+} from "./pane";
+export { renderPane } from "./render";
+export { updateBubbles, resolveStopName } from "./bubbles";
+export { pushDecision } from "./decision";
+export { updateModeBadge } from "./mode-badge";
